third_party/blink/web_tests/display-lock/activation/focus-updated-style.html

<!doctype HTML>

<!--
Focus on a div that has updated style/layout.
-->

<div id="container" style ="contain:style layout">
  <div id="focusable" tabIndex="0">
    focusable thing
  </div>
</div>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
promise_test((t) => {
  const acquirePromise = container.displayLock.acquire({ timeout: Infinity, activatable: true });
  return new Promise((resolve, reject) => {
    acquirePromise.then(() => {
      container.onbeforeactivate = reject;
      focusable.style = "display: none";
      focusable.focus();
      t.step(() => assert_not_equals(document.activeElement, focusable));
      resolve();
    });
  });
}, "Trying to focus on an element in a locked subtree that's not visible anymore won't work");
</script>
